Fix test Meekmock
